Member Test Code complete
이제 팀테이블의 유저를 삭제하면 삭제 요청한 학생의 이메일로 대체하는 코드만 추가하면 되지롱

+++ src/main/java/com/dsmpear/main/service/member/MemberServiceImpl.java
@@ -6,10 +6,7 @@ import com.dsmpear.main.entity.team.Team;
 import com.dsmpear.main.entity.team.TeamRepository;
 import com.dsmpear.main.entity.user.User;
 import com.dsmpear.main.entity.user.UserRepository;
-import com.dsmpear.main.exceptions.MemberAlreadyIncludeException;
-import com.dsmpear.main.exceptions.TeamNotFoundException;
-import com.dsmpear.main.exceptions.UserNotFoundException;
-import com.dsmpear.main.exceptions.UserNotMemberException;
+import com.dsmpear.main.exceptions.*;
 import com.dsmpear.main.payload.request.MemberRequest;
 import com.dsmpear.main.security.auth.AuthenticationFacade;
 import lombok.RequiredArgsConstructor;
@@ -17,9 +14,10 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MemberServiceImpl implements MemberService{
+public class MemberServiceImpl implements MemberService {
 
     private final AuthenticationFacade authenticationFacade;
+
     private final UserRepository userRepository;
     private final MemberRepository memberRepository;
     private final TeamRepository teamRepository;
@@ -30,42 +28,50 @@ public class MemberServiceImpl implements MemberService{
         User user=userRepository.findByEmail(authenticationFacade.getUserEmail())
                 .orElseThrow(UserNotFoundException::new);
 
-        //요청한 user가 팀 멤버인지 확인하기
-        Member member=memberRepository.findByUserEmail(memberRequest.getUserEmail())
-                .orElseThrow(UserNotMemberException::new);
+        userRepository.findByEmail(memberRequest.getUserEmail())
+                .orElseThrow(UserNotFoundException::new);
 
         //멤버 이용해서 팀 찾기
-        Team team=teamRepository.findById(member.getTeamId())
-                .filter(t -> user.getEmail().equals(t.getUserEmail()))
+        Team team=teamRepository.findById(memberRequest.getTeamId())
                 .orElseThrow(TeamNotFoundException::new);
 
+        //요청한 user가 팀 멤버인지 확인하기
+        memberRepository.findByTeamIdAndUserEmail(team.getId(), user.getEmail())
+                .orElseThrow(UserNotMemberException::new);
+
         //팀 아이디랑 유저 이메일로 찾기
-        memberRepository.findByTeamIdAndUserEmail(team.getId(),user.getEmail())
-                .ifPresent(m -> {throw new MemberAlreadyIncludeException();
+        memberRepository.findByTeamIdAndUserEmail(team.getId(), memberRequest.getUserEmail())
+                .ifPresent(m -> {
+                    throw new MemberAlreadyIncludeException();
                 });
 
         memberRepository.save(
                 Member.builder()
-                        .teamId(team.getId())
-                        .userEmail(user.getEmail())
-                .build()
+                    .teamId(team.getId())
+                    .userEmail(memberRequest.getUserEmail())
+                    .build()
         );
     }
 
     @Override
-    public void deleteMember(String userEmail) {
+    public void deleteMember(Integer memberId) {
         //요청한 user가 팀 멤버인지 확인하기
         User user=userRepository.findByEmail(authenticationFacade.getUserEmail())
                 .orElseThrow(UserNotFoundException::new);
 
-        Member member=memberRepository.findByUserEmail(userEmail)
-                .orElseThrow(UserNotMemberException::new);
-
-        teamRepository.findById(member.getTeamId())
-                .filter(team -> user.getEmail().equals(member.getUserEmail()))
+        Team team=teamRepository.findById(memberId)
                 .orElseThrow(TeamNotFoundException::new);
 
-        memberRepository.delete(member);
+        Member member=memberRepository.findById(memberId)
+                .orElseThrow(UserNotMemberException::new);
 
+        if(user.getEmail().equals(member.getUserEmail())){
+            throw new UserEqualsMemberException();
+        }
+
+        //만약 최초 생성자를 지우려고 하면, 삭제를 요청한 유저 이메일로 대체하기
+
+        memberRepository.delete(member);
     }
+
 }
